requirements.txt
mysql-connector-python
halo